+++ src/front/LogIn.java
@@ -3,6 +3,7 @@ package front;
 import back.ResponseCode;
 import back.request.account.Login_Request;
 import back.response.account.Login_Response;
+import front.MainPage.MainPage;
 
 import javax.swing.*;
 import java.awt.*;
